backend/requirements.txt
Django>=3.0,<4.0
psycopg2==2.9.5
djangorestframework
markdown
django-filter
django-admin-rangefilter
django-admin-numeric-filter
djangorestframework-simplejwt
django-cors-headers
django-rest-swagger